val.criterion Options

Internally used, package-specific options. All options will prioritize R options() values, and fall back to environment variables if undefined. If neither the option nor the environment variable is set, a default value is used.

Checking Option Values

Option values specific to val.criterion can be accessed by passing the package name to env.

options::opts(env = "val.criterion")

options::opt(x, default, env = "val.criterion")

Options

repos

Metrics repositories to use implicitly for package filters. Will be accessed using available.packages(), so repository paths should follow similar conventions.
default:

character(0L)

option:

val.criterion.repos

envvar:

R_VAL_CRITERION_REPOS (evaluated if possible, raw string otherwise)

exceptions

Packages that should be treated as exceptions. Even if they would otherwise be filtered out using preset filters, they would still be discoverable and installable without issue.
default:

character(0L)

option:

val.criterion.exceptions

envvar:

R_VAL_CRITERION_EXCEPTIONS (evaluated if possible, raw string otherwise)

quiet

Wether filtering operations should be reported as they're being queried. Passed to the quiet parameter of individual functions where it's used.
default:

FALSE

option:

val.criterion.quiet

envvar:

R_VAL_CRITERION_QUIET (evaluated if possible, raw string otherwise)

actions

A set of calls that should prompt actions. See default_actions for an example of the expected format.
default:

default_actions()

option:

val.criterion.actions

envvar:

R_VAL_CRITERION_ACTIONS (evaluated if possible, raw string otherwise)

See also